invalid write of size 1 solving

+++ WORKSHOP - 02/part 1/TennisLog.h
@@ -53,7 +53,7 @@ namespace sdds {
 		//a dynamic array of TennisMatches
 		TennisMatch* m_tennisMatch;
 		//a count of how many are being stored
-		int m_storedMatches{};
+		size_t m_storedMatches{};
 
 	public:
 		// Default constructor
